+++ arrow-validation/gradle.properties
@@ -1,4 +1,4 @@
 # Maven publishing configuration
 POM_NAME=Validation Module
-POM_ARTIFACT_ID=funktionale-validation
+POM_ARTIFACT_ID=arrow-validation
 POM_PACKAGING=jar
